Add flattener for dictionary objects
Values are supposed to be stored in form of key-value pairs with
keys being paths of the original JSON object. In order to convert
JSON objects to the key-value representation, they will be get as
python dictionary. This dictionary has to be flatten so that it
only contains primitive values.

The flattener is a class that flattens the dictionary in the described
way.

Unit tests included.